+++ hummingbot/connector/exchange/bitcoin_rd/bitcoin_rd_auth.py
@@ -10,7 +10,7 @@ from hummingbot.core.web_assistant.auth import AuthBase
 from hummingbot.core.web_assistant.connections.data_types import RESTMethod, RESTRequest, WSRequest
 
 
-class BinanceAuth(AuthBase):
+class BitcoinRDAuth(AuthBase):
     def __init__(self, api_key: str, secret_key: str, time_provider: TimeSynchronizer):
         self.api_key = api_key
         self.secret_key = secret_key
@@ -37,7 +37,7 @@ class BinanceAuth(AuthBase):
 
     async def ws_authenticate(self, request: WSRequest) -> WSRequest:
         """
-        This method is intended to configure a websocket request to be authenticated. Binance does not use this
+        This method is intended to configure a websocket request to be authenticated. BitcoinRD does not use this
         functionality
         """
         return request  # pass-through
